requirements.txt
alabaster==0.7.10
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
coveralls==1.1
docopt==0.6.2
docutils==0.13.1
greenlet==0.4.12
hypothesis==3.14.0
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
msgpack-python==0.4.8
nose==1.3.7
Pint==0.8.1
Pygments==2.2.0
pyserial==3.4
pytz==2017.2
requests==2.18.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
urllib3==1.22
